finish FIXME: move open_browser() to addon/browser
